+++ introspect/functions.py
@@ -5142,27 +5142,6 @@ FUNCTIONS: Mapping[str, FunctionDecl] = dict([
          ),
          doc='Write [datetime, type: message] to MUJOCO_LOG.TXT.',
      )),
-    ('mj_activate',
-     FunctionDecl(
-         name='mj_activate',
-         return_type=ValueType(name='int'),
-         parameters=(
-             FunctionParameterDecl(
-                 name='filename',
-                 type=PointerType(
-                     inner_type=ValueType(name='char', is_const=True),
-                 ),
-             ),
-         ),
-         doc='Return 1 (for backward compatibility).',
-     )),
-    ('mj_deactivate',
-     FunctionDecl(
-         name='mj_deactivate',
-         return_type=ValueType(name='void'),
-         parameters=(),
-         doc='Do nothing (for backward compatibility).',
-     )),
     ('mju_zero3',
      FunctionDecl(
          name='mju_zero3',